API Reference

Constants

Tempo.DJ2000 — Constant

DJ2000

Reference epoch (J2000.0), Julian Date (2451545.0). It is 12:00 01-01-2000.

Tempo.DJM0 — Constant

DJM0

Julian Date of Modified Julian Date zero point (2400000.5). It is 00:00 17-11-1858.

Tempo.DMJD — Constant

DMJD

Reference epoch (J2000.0), Modified Julian Date (51544.5).

Tempo.GPS — Constant

GPS

The singleton instance of the GlobalPositioningSystemTime type representing the Global Positioning System Time (GPS) time scale.

Tempo.J2000 — Constant

J2000

The singleton instance of the JulianDate2000 epoch origin.

With this origin, Epoch reference is 2000-01-01T12:00.

Tempo.JD — Constant

JD

The singleton instance of the JulianDate epoch origin.

With this origin, Epoch reference is -4712-01-01T12:00.

Tempo.LEAPSECONDS — Constant

LEAPSECONDS

Leapseconds data.

Tempo.MJ2000 — Constant

MJ2000

The singleton instance of the ModifiedJulianDate2000 epoch origin.

With this origin, Epoch reference is 2000-01-01T00:00.

Tempo.MJD — Constant

MJD

The singleton instance of the ModifiedJulianDate epoch origin.

With this origin, Epoch reference is 1858-11-17T00:00.

Tempo.TAI — Constant

TAI

The singleton instance of the InternationalAtomicTime type representing the International Atomic Time (TAI) time scale.

Tempo.TCB — Constant

TCB

The singleton instance of the BarycentricCoordinateTime type representing the Barycentric Coordinate Time (TCB) time scale.

Tempo.TCG — Constant

TCG

The singleton instance of the GeocentricCoordinateTime type representing the Geocentric Coordinate Time (TCG) time scale.

Tempo.TDB — Constant

TDB

The singleton instance of the BarycentricDynamicalTime type representing the Barycentric Dynamical Time (TDB) time scale.

Tempo.TDBH — Constant

TDBH

The singleton instance of the HighPrecisionBarycentricDynamicalTime type representing the High Precision Barycentric Dynamical Time (TDBH) time scale.

Tempo.TIMESCALES — Constant

TIMESCALES

Time scales graph, containing at least: ("TT", "TAI", "UTC", "TCG", "TCB", "TDB", "UT1", "TDBH", "GPS")

It can be easily extended using the @timescale to create new TimeScaleNode aliases and add_timescale method to define its relation with the other nodes in the graph.

Example

# define a new timescale type alias
@timescale NTS 100 NewTimeScale

# define offset to and from another timescale in the graph 
offset_ffp(seconds) = 1.0
offset_ftp(seconds) = -1.0

# connect to the graph, with the parent node (TDB in this example)
add_timescale(TIMESCALES, NTS, offset_ffp, parent=TDB, ftp=offset_ftp)

Tempo.TT — Constant

TT

The singleton instance of the TerrestrialTime type representing the Terrestrial Time (TT) time scale.

Tempo.UT1 — Constant

UT1

The singleton instance of the UniversalTime type representing the Universal Time (UT1) time scale.

Tempo.UTC — Constant

UTC

The singleton instance of the CoordinatedUniversalTime type representing the Coordinated Universal Time (UTC) time scale.

Types

Tempo.AbstractEpochOrigin — Type

AbstractEpochOrigin

Abstract type for all epoch origins.

Tempo.AbstractTimeScale — Type

AbstractTimeScale

All timescales are subtypes of the abstract type AbstractTimeScale.

Tempo.BarycentricCoordinateTime — Type

BarycentricCoordinateTime

A type representing the Barycentric Coordinate Time (TCB) time scale.

Tempo.BarycentricDynamicalTime — Type

BarycentricDynamicalTime

A type representing the Barycentric Dynamical Time (TDB) time scale.

Tempo.CoordinatedUniversalTime — Type

CoordinatedUniversalTime

A type representing the Coordinated Universal Time (UTC) time scale.

Tempo.Date — Type

Date

Type to represent a calendar date.

Fields

year – year

month – month

day – day

Constructors

Date(year::N, month::N, day::N) – is the default constructor.

Date(offset::N) where {N<:Integer} – initialize from integer day offset from 2000-01-01.

Date(d::Date, offset::N) where {N<:Integer} – day offset from d.

Date(year::N, dayinyear::N) where {N<:Integer} – initialize giving the year and the day of the year.

Date(dt::DateTime) – extract date from DateTime objects.

Tempo.DateTime — Type

DateTime{N<:Integer, T<:AbstractFloat} <: AbstractDateTimeEpoch

A type wrapping a date and a time since a reference date.

Fields

date – Date part of the type

time – Time part of the type

Constructors

DateTime{T}(date::Date{N}, time::Time{N, T}) – default constructor.

DateTime(year::N, month::N, day::N, hour::N, min::N, sec::N, frac::T=0.0) where {N<:Integer, T<:AbstractFloat} – full constructor

DateTime(s::AbstractString) – parse DateTime from ISO string using parse_iso.

DateTime(seconds::T) where {T<:AbstractFloat} – parse from seconds since J2000.

DateTime(d::Date, sec::T) where {T<:AbstractFloat} – parse as seconds since d.

DateTime{N, T}(dt::DateTime) where {N, T} – ghost constructor

DateTime(e::Epoch) – construct from Epoch

Tempo.Epoch — Type

Epoch

A type to represent Epoch-like data. Epochs here are represented as seconds + fraction since a reference epoch, which is considered to be 2000-01-01T12:00:00, i.e. J2000.

Fields

scale – TimeScale to represent the date

seconds – seconds since origin

Constructors

Epoch{S}(seconds::T) where {S<:AbstractTimeScale, T<:AbstractFloat} – default

Epoch(seconds::T, ::S) where {S<:AbstractTimeScale, T<:AbstractFloat}

Epoch(dt::DateTime, ::S) where {S<:AbstractTimeScale} – construct from DateTime

Tempo.Epoch — Method

Epoch(s::AbstractString, scale::AbstractTimeScale)

Construct an Epoch from a str in the AbstractTimeScale (scale).

Note

This constructor requires that the str is in the format yyyy-mm-ddTHH:MM:SS.sss.

Tempo.Epoch — Method

Epoch(s::AbstractString)

Construct an Epoch from a str.

Note

This constructor requires that the str is in the format:

ISO – yyyy-mm-ddTHH:MM:SS.ffff : assume J2000 as origin

J2000 – DDDD.ffff : parse Julian Dates since J2000 (days)

JD – JD DDDDDDDDD.ffffff : parse Julian Date (days)

MJD – MJD DDDDDDDDD.ffffff : parse Julian Dates since (days)

In all cases the TimeScale should be added at the end of the string, separated by a whitespace. If it is not declared, TDB will be used as timescale.

Examples

julia> # standard ISO string 
julia> Epoch("2050-01-01T12:35:15.0000 TT")
2050-01-01T12:35:14.9999 TT

julia> # standard ISO string (without scale)
julia> Epoch("2050-01-01T12:35:15.0000")
2050-01-01T12:35:14.9999 TDB

julia> # Parse Julian Dates 
julia> Epoch("JD 2400000.5")
1858-11-17T00:00:00.0000 TDB

julia> # Parse Modified Julian Dates
julia> Epoch("MJD 51544.5")
2000-01-01T12:00:00.0000 TDB

julia> # Parse Julian Dates since J2000 
julia> Epoch("12.0")
2000-01-13T12:00:00.0000 TDB

julia> # All Julian Date parsers allow timescales 
julia> Epoch("12.0 TT")
2000-01-13T12:00:00.0000 TT

Tempo.EpochConversionError — Type

true

A type representing epoch conversion errors.

Tempo.GeocentricCoordinateTime — Type

GeocentricCoordinateTime

A type representing the Geocentric Coordinate Time (TCG) time scale.

Tempo.GlobalPositioningSystemTime — Type

GlobalPositioningSystemTime

A type representing the Global Positioning System Time (GPS) time scale.

Tempo.HighPrecisionBarycentricDynamicalTime — Type

HighPrecisionBarycentricDynamicalTime

A type representing the High Precision Barycentric Dynamical Time (TDBH) time scale.

Tempo.InternationalAtomicTime — Type

InternationalAtomicTime

A type representing the International Atomic Time (TAI) time scale.

Tempo.JulianDate — Type

JulianDate

A type representing the JulianDate (JD) epoch origin.

With this origin, Epoch reference is -4712-01-01T12:00.

Tempo.JulianDate2000 — Type

JulianDate2000

A type representing the JulianDate2000 (J2000) epoch origin.

With this origin, Epoch reference is 2000-01-01T12:00.

Tempo.Leapseconds — Type

Leapseconds{T}

Stores information about leap seconds that have been added to Coordinated Universal Time (UTC) since the start of the year 2000.

Fields

lastupdate: a DatesDateTime object representing the date and time when the Leapseconds struct was last updated.

jd2000: a vector storing the Julian Date, in days since J2000, of each leap second.

leap: a vector storing the number of leap seconds at each corresponding entry of the jd2000 field.

Tempo.ModifiedJulianDate — Type

ModifiedJulianDate

A type representing the ModifiedJulianDate (MJD) epoch origin.

With this origin, Epoch reference is 1858-11-17T00:00.

Tempo.ModifiedJulianDate2000 — Type

ModifiedJulianDate2000

A type representing the ModifiedJulianDate2000 (MJ2000) epoch origin.

With this origin, Epoch reference is 2000-01-01T00:00.

Tempo.TerrestrialTime — Type

TerrestrialTime

A type representing the Terrestrial Time (TT) time scale.

Tempo.Time — Type

Time{T<:AbstractFloat}

A type representing the time of the day.

Fields

hour – hour

minute – minute

second – second

fraction – fraction of seconds

Constructors

Time{T}(hour::N, minute::N, second::N, fraction::T) where {N<:Integer, T<:AbstractFloat} – default constructor.

Time(hour::N, minute::N, second::T) where {N<:Integer, T<:AbstractFloat} – initialize a Time type automatically computing seconds fractions

Time(secondinday::Integer, fraction::T) where {T<:AbstractFloat} – type from second in the day and fraction of seconds.

Time(dt::DateTime) – extract time from DateTime objects.

Tempo.UniversalTime — Type

UniversalTime

A type representing the Universal Time (UT1) time scale.

Types

Base.convert — Method

convert(to::S2, e::Epoch{S1}; system::TimeSystem=TIMESCALES)

Convert Epoch with timescale S1 to S2. Allows to use the default TimeSystem or a custom constructed one.

Graphs.SimpleGraphs.add_vertex! — Method

add_vertex!(s::TimeSystem, ts::TimeScaleNode)

Register a new node in the TimeSystem.

Tempo.cal2jd — Method

cal2jd(d::Date)

Convert Gregorian calendar Date to a Julian Date, in days.

Outputs

j2000 – J2000 zero point: always 2451545

d – J2000 Date for 12 hrs

Tempo.cal2jd — Method

cal2jd(year::Integer, month::Integer, day::Integer)

This function converts a given date in the Gregorian calendar (year, month, day) to the corresponding two-parts Julian Date. The first part is the DJ2000, while the second output is the number of days since DJ2000.

The year must be greater than 1583, and the month must be between 1 and 12. The day must also be valid, taking into account whether the year is a leap year. If the input year or month or day are invalid, a DomainError is thrown.

Examples

julia> cal2jd(2021, 1, 1)
(2.4000005e6, 59215.0)

julia> cal2jd(2022, 2, 28)
(2.4000005e6, 59638.0)

#TODO: add example for epoch conversion error

References

Seidelmann P. K., (1992), Explanatory Supplement to the Astronomical Almanac, University Science Books, Section 12.92 (p604).

Klein, A., (2006), A Generalized Kahan-Babuska-Summation-Algorithm. Computing, 76, 279-293, Section 3.

ERFA software library

Tempo.calhms2jd — Method

calhms2jd(year, month, day, hour, minute, seconds)

Convert Gregorian Calendar date and time to a two-parts Julian Date. The first part is the DJ2000, while the second output is the number of days since DJ2000.

Tempo.day — Method

day(d::DateTime)

Get day associated to a DateTime type.

Tempo.day — Method

day(d::Date)

Get day associated to a Date d.

Tempo.fd2hms — Method

fd2hms(fd::Number)

Convert the day fraction fd to hour, minute and seconds.

Tempo.fd2hmsf — Method

fd2hmsf(fd::Number)

Convert the day fraction fd to hour, minute, second and fraction of seconds.

Tempo.find_day — Method

find_day(dayinyear::Integer, month::Integer, isleap::Bool)

Find day from the day in the year, the month and using if the year is leap or not.

Tempo.find_dayinyear — Method

find_dayinyear(d::Date)

Find the day in the year.

Tempo.find_dayinyear — Method

find_dayinyear(month::Integer, day::Integer, isleap::Bool)

Find the day of the year given the month, the day of the month and whether the year is leap or not.

Tempo.find_month — Method

find_month(dayinyear::Integer, isleap::Bool)

Find the month from the day of the year, depending on whether the year is leap or not.

Tempo.find_year — Method

find_year(d::Integer)

Return the Gregorian year associated to the given Julian Date day d.

Tempo.fraction_of_day — Method

fraction_of_day(t::Time)
hms2fd(t::Time)

Find fraction of day.

Example

julia> t = Time(12, 30, 40.3424)
12:30:40.3423
julia> Tempo.fraction_of_day(t)
0.5213002592592593  # days

Tempo.fraction_of_second — Method

fraction_of_second(t::Time)

Find fraction of seconds.

Example

julia> t = Time(12, 30, 40.3424)
12:30:40.3423
julia> Tempo.fraction_of_second(t)
0.3423999999999978  # seconds

Tempo.get_leapseconds — Method

get_leapseconds()

Parse leapseconds data and return a Leapseconds type.

Note

The leapsecond kernel is retrieved from the artifacts of this package. This artifact will be updated whenever a new leapsecond is added.

Tempo.hms2fd — Method

hms2fd(h::Integer, m::Integer, s::Number)

Convert hours, minutes and seconds to day fraction.

Tempo.hour — Method

hour(d::DateTime)

Get hour associated to a DateTime type.

Tempo.hour — Method

hour(t::Time)

Get the current hour.

Tempo.isleapyear — Method

isleapyear(d::Date)

True if Date is within a leap year.

Tempo.isleapyear — Method

isleapyear(year::Integer)

Return true if the given Gregorian year is leap.

Tempo.j2000 — Method

j2000(jd)
j2000(jd1, jd2)

Convert Julian Date in days since J2000

Tempo.j2000 — Method

j2000(dt::DateTime)

Convert a DateTime dt in Julian days since J2000

Tempo.j2000 — Method

j2000(d::Date)

Convert Gregorian calendar date Date to Julian Date since DJ2000, in days.

Tempo.j2000 — Method

j2000(e::Epoch)

Convert Epoch in Julian Date since J2000 (days)

Tempo.j2000c — Method

j2000c(jd)
j2000c(jd1, jd2)

Convert Julian Date (in days) to Julian Centuries

Tempo.j2000c — Method

j2000c(dt::DateTime)

Convert a DateTime dt in a Julian Date since DJ2000, in centuries.

Tempo.j2000c — Method

j2000c(e::Epoch)

Convert Epoch in Julian Date since J2000 (centuries)

Tempo.j2000s — Method

j2000s(jd)
j2000s(jd1, jd2)

Convert Julian Date (in days) in seconds past J2000

Tempo.j2000s — Method

j2000s(dt::DateTime)

Convert a DateTime dt to seconds since J2000

Tempo.j2000s — Method

j2000s(e::Epoch)

Convert Epoch in Julian Date since J2000 (seconds)

Tempo.jd2cal — Method

jd2cal(dj1::Number, dj2::Number)

This function converts a given Julian Date (JD) to a Gregorian calendar date (year, month, day, and fraction of a day).

The JD is composed of two input arguments, dj1 and dj2, which can be apportioned in any convenient way. For example, JD 2450123.7 could be expressed as dj1 = 2450123.7 and dj2 = 0.0, or as dj1 = 2451545.0 and dj2 = -1421.3, or in any of the other ways listed in the documentation.

The earliest valid date that can be converted is -4713 Jan 1, and the largest value accepted is 1e9. If the input JD is outside this range, an EpochConversionError is thrown.

Note

The Julian Date is apportioned in any convenient way between the arguments dj1 and dj2. For example, JD = 2450123.7 could be expressed in any of these ways, among others:

dj1	dj2

2450123.7	0.0	(JD method)

2451545.0	-1421.3	(J2000 method)

2400000.5	50123.2	(MJD method)

2450123.5	0.2	(date & time method)

Warning

The earliest valid date is 0 (-4713 Jan 1). The largest value accepted is 1e9.

References

Seidelmann P. K., (1992), Explanatory Supplement to the Astronomical Almanac, University Science Books, Section 12.92 (p604).

Klein, A., (2006), A Generalized Kahan-Babuska-Summation-Algorithm. Computing, 76, 279-293, Section 3.

ERFA software library

Tempo.jd2calhms — Method

jd2calhms(dj1::Number, dj2::Number)

Convert a two-parts Julian Date to Gregorian year, month, day, hour, minute, seconds. See jd2cal for more information on the Julian Date composition.

Tempo.leapseconds — Method

leapseconds(jd2000::Number)

For a given UTC date, in Julian days since J2000, calculate Delta(AT) = TAI - UTC.

Tempo.microsecond — Method

microsecond(t::Time)

Get the current microsecond.

Tempo.millisecond — Method

millisecond(t::Time)

Get the current millisecond.

Tempo.minute — Method

minute(d::DateTime)

Get minute associated to a DateTime type.

Tempo.minute — Method

minute(t::Time)

Get the current minute.

Tempo.month — Method

month(d::DateTime)

Get month associated to a DateTime type.

Tempo.month — Method

month(d::Date)

Get month associated to a Date d.

Tempo.nanosecond — Method

nanosecond(t::Time)

Get the current nanosecond.

Tempo.offset — Method

offset(::JulianDate2000)

Offset in days to shift J2000 epochs (with origin at 2000-01-01T12:00) to JulianDate2000 (with origin at 2000-01-01T12:00)

Tempo.offset — Method

offset(::JulianDate)

Offset in days to shift J2000 epochs (with origin at 2000-01-01T12:00) to JulianDate (with origin at -4712-01-01T12:00)

Tempo.offset — Method

offset(::ModifiedJulianDate2000)

Offset in days to shift J2000 epochs (with origin at 2000-01-01T12:00) to ModifiedJulianDate2000 (with origin at 2000-01-01T00:00)

Tempo.offset — Method

offset(::ModifiedJulianDate)

Offset in days to shift J2000 epochs (with origin at 2000-01-01T12:00) to ModifiedJulianDate (with origin at 1858-11-17T00:00)

Tempo.offset_gps2tai — Method

offset_gps2tai(seconds)

Return the fixed offset between GPS and TAI in seconds.

Tempo.offset_tai2gps — Method

offset_tai2gps(seconds)

Return the fixed offset between TAI and GPS in seconds.

Tempo.offset_tai2tt — Method

offset_tai2tt(seconds)

Return the fixed offset between TAI and TT in seconds.

Tempo.offset_tai2utc — Method

offset_tai2utc(seconds)

Return the offset between TAI and UTC in seconds.

Tempo.offset_tcb2tdb — Method

offset_tcb2tdb(seconds)

Return the linear offset between TCB and TDB in seconds.

Tempo.offset_tcg2tt — Method

offset_tcg2tt(seconds)

Return the linear offset between TCG and TT in seconds.

Tempo.offset_tdb2tcb — Method

offset_tdb2tcb(seconds)

Return the linear offset between TDB and TCB in seconds.

Tempo.offset_tdb2tt — Method

offset_tdb2tt(seconds)

Return the offset between TDB and TT in seconds. This routine is accurate to ~40 microseconds over the interval 1900-2100.

Note

An accurate transformation between TDB and TT depends on the trajectory of the observer. For two observers fixed on Earth's surface the quantity TDB-TT can differ by as much as ~4 microseconds.

References

https://www.cv.nrao.edu/~rfisher/Ephemerides/times.html#TDB

Issue #26

Tempo.offset_tt2tai — Method

offset_tt2tai(seconds)

Return the fixed offset between TT and TAI in seconds.

Tempo.offset_tt2tcg — Method

offset_tt2tcg(seconds)

Return the linear offset between TT and TCG in seconds.

Tempo.offset_tt2tdb — Method

offset_tt2tdb(seconds)

Return the offset between TT and TDB in seconds. This routine is accurate to ~40 microseconds over the interval 1900-2100.

Note

An accurate transformation between TDB and TT depends on the trajectory of the observer. For two observers fixed on Earth's surface the quantity TDB-TT can differ by as much as ~4 microseconds.

References

https://www.cv.nrao.edu/~rfisher/Ephemerides/times.html#TDB

Issue #26

Tempo.offset_tt2tdbh — Method

offset_tt2tdbh(seconds)

Return the offset between TT and TDBH in seconds.

The maximum error in using the above formula is about 10 µs from 1600 to 2200. For even more precise applications, the series expansion by Harada & Fukushima (2003) is recommended.

References

The IAU Resolutions on Astronomical Reference Systems, Time Scales, and Earth Rotation Models, United States Naval Observatory, https://arxiv.org/pdf/astro-ph/0602086.pdf

Tempo.offset_utc2tai — Method

offset_utc2tai(seconds)

Return the offset between UTC and TAI in seconds.

Tempo.parse_iso — Method

parse_iso(s::AbstractString)

Parse an ISO datetime string, in the format YYYY-MM-DDThh:mm:ss.ffffffff and return a Tuple containing the year, month, day , hour, minute, second and milliseconds. If the string format is not recognised as ISO, an ArgumentError is thrown.

Tempo.second — Method

second(d::DateTime)

Get the seconds associated to a DateTime type.

Tempo.second — Method

second(::Type{<:AbstractFloat}, t::Time)
second(::Type{<:Integer}, t::Time)
second(t::Time)

Get the current second.

Tempo.second_in_day — Method

second_in_day(t::Time)::AbstractFloat

Find second in the day.

Example

julia> t = Time(12, 30, 40.3424)
12:30:40.3423
julia> Tempo.second_in_day(t)
45040.3424  # seconds

Tempo.tai2utc — Method

tai2utc(tai1, tai2)

Transform a 2-part (quasi) Julian Date, in days, in International Atomic Time, TAI to a 2-part Julian Date in the Coordinated Universal Time, UTC, scale.

Note

tai1 + tai2 is Julian Date, apportioned in any convenient way between the two arguments, for example such that tai1 is the Julian Day Number and tai2 is the fraction of a day. The returned utc1 and utc2 form an analogous pair.

Note

JD cannot unambiguously represent UTC during a leap second unless special measures are taken. The convention in the present function is that the JD day represents UTC days whether the length is 86399, 86400 or 86401 SI seconds.

References

Seidelmann P. K., (1992), Explanatory Supplement to the Astronomical Almanac, University Science Books, Section 12.92 (p604).

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

ERFA software library

Tempo.timescale — Method

timescale(ep::Epoch)

Epoch timescale.

Tempo.utc2tai — Method

utc2tai(utc1, utc2)

Transform a 2-part (quasi) Julian Date, in days, in Coordinate Universal Time, UTC to a 2-part Julian Date in the International Atomic Time, TAI scale.

Note

utc1 + utc2 is quasi Julian Date (see Note 2), apportioned in any convenient way between the two arguments, for example such that utc1 is the Julian Day Number and utc2 is the fraction of a day.

Note

JD cannot unambiguously represent UTC during a leap second unless special measures are taken. The convention in the present function is that the JD day represents UTC days whether the length is 86399, 86400 or 86401 SI seconds.

References

Seidelmann P. K., (1992), Explanatory Supplement to the Astronomical Almanac, University Science Books, Section 12.92 (p604).

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

ERFA software library

Tempo.value — Method

value(ep::Epoch)

Full Epoch value.

Tempo.year — Method

year(d::DateTime)

Get year associated to a DateTime type.

Tempo.year — Method

year(d::Date)

Get year associated to a Date d.

Macros

Tempo.@timescale — Macro

@timescale(name, id, type)

Create an AbstractTimeScale instance to alias the given id.

Inputs

name – Acronym to denote the new time scale

id – Integer whose alias is the new time scale

type – Name of the type representing the new time scale