API Reference

Constants

Tempo.DJ2000 — Constant

DJ2000

Reference epoch (J2000.0), Julian Date (2451545.0). It is 12:00 01-01-2000.

Tempo.DJM0 — Constant

DJM0

Julian Date of Modified Julian Date zero point (2400000.5). It is 00:00 17-11-1858.

Tempo.DMJD — Constant

DMJD

Reference epoch (J2000.0), Modified Julian Date (51544.5).

Tempo.GPS — Constant

GPS

The singleton instance of the GlobalPositioningSystemTime type representing the Global Positioning System Time (GPS) time scale.

Tempo.J2000 — Constant

J2000

The singleton instance of the JulianDate2000 epoch origin.

With this origin, Epoch reference is 2000-01-01T12:00.

Tempo.JD — Constant

JD

The singleton instance of the JulianDate epoch origin.

With this origin, Epoch reference is -4712-01-01T12:00.

Tempo.LEAPSECONDS — Constant

LEAPSECONDS

Leapseconds data.

Tempo.MJD — Constant

MJD

The singleton instance of the ModifiedJulianDate epoch origin.

With this origin, Epoch reference is 1858-11-17T00:00.

Tempo.MJD2000 — Constant

MJD2000

The singleton instance of the ModifiedJulianDate2000 epoch origin.

With this origin, Epoch reference is 2000-01-01T00:00.

Tempo.TAI — Constant

TAI

The singleton instance of the InternationalAtomicTime type representing the International Atomic Time (TAI) time scale.

Tempo.TCB — Constant

TCB

The singleton instance of the BarycentricCoordinateTime type representing the Barycentric Coordinate Time (TCB) time scale.

Tempo.TCG — Constant

TCG

The singleton instance of the GeocentricCoordinateTime type representing the Geocentric Coordinate Time (TCG) time scale.

Tempo.TDB — Constant

TDB

The singleton instance of the BarycentricDynamicalTime type representing the Barycentric Dynamical Time (TDB) time scale.

Tempo.TDBH — Constant

TDBH

The singleton instance of the HighPrecisionBarycentricDynamicalTime type representing the High Precision Barycentric Dynamical Time (TDBH) time scale.

Tempo.TIMESCALES — Constant

TIMESCALES

Default time scales graph, containing at least: ("TT", "TAI", "UTC", "TCG", "TCB", "TDB", "UT1", "TDBH", "GPS")

It can be easily extended using the @timescale to create new TimeScaleNode aliases and add_timescale! method to define its relation with the other nodes in the graph.

Example

# define a new timescale type alias
@timescale NTS 100 NewTimeScale

# define offset to and from another timescale in the graph 
offset_ffp(seconds) = 1.0
offset_ftp(seconds) = -1.0

# connect to the graph, with the parent node (TDB in this example)
add_timescale!(TIMESCALES, NTS, offset_ffp, parent=TDB, ftp=offset_ftp)

### See also 
See also [`@timescale`](@ref), [`TimeScaleNode`](@ref) and [`add_timescale`](@ref).

Tempo.TT — Constant

TT

The singleton instance of the TerrestrialTime type representing the Terrestrial Time (TT) time scale.

Tempo.UT1 — Constant

UT1

The singleton instance of the UniversalTime type representing the Universal Time (UT1) time scale.

Tempo.UTC — Constant

UTC

The singleton instance of the CoordinatedUniversalTime type representing the Coordinated Universal Time (UTC) time scale.

Types

Tempo.AbstractEpochOrigin — Type

AbstractEpochOrigin

Abstract type for all epoch origins.

Tempo.AbstractTimeScale — Type

AbstractTimeScale

All timescales are subtypes of the abstract type AbstractTimeScale.

Tempo.BarycentricCoordinateTime — Type

BarycentricCoordinateTime <: AbstractTimeScale

A type representing the Barycentric Coordinate Time (TCB) time scale.

Tempo.BarycentricDynamicalTime — Type

BarycentricDynamicalTime <: AbstractTimeScale

A type representing the Barycentric Dynamical Time (TDB) time scale.

Tempo.CoordinatedUniversalTime — Type

CoordinatedUniversalTime <: AbstractTimeScale

A type representing the Coordinated Universal Time (UTC) time scale.

Tempo.Date — Type

Date

A type to represent a calendar date by storing the year, month and day.

Date(year::Int, month::Int, day::Int)

Construct a Date object given the year, month and day.

Date(offset::Integer)

Create a Date object given an integer number of days since 2000-01-01.

Date(year::Integer, dayinyear::Integer)

Create a Date object given the year and the day of the year dayinyear.

Examples

julia> Date(2020, 1)
2020-01-01

julia> Date(2020, 300)
2020-10-26

Date(dt::DateTime)

Extract the Date object from a DateTime structure.

See also

See also Time and DateTime.

Tempo.DateTime — Type

DateTime{T}

A type wrapping a Date and a Time object.

DateTime(date::Date, time::Time{T})

Construct a DateTime object of type T from its Date and Time components.

DateTime(year::Int, month::Int, day::Int, hour::Int, min::Int, sec::Int, frac::Number)

Create a DateTime object by parts.

DateTime(iso::AbstractString)

Create a DateTime object from by parsing an ISO datetime string iso, in the format YYYY-MM-DDThh:mm:ss.ffffffff. The DateTime parts not provided in the string will be assigned default values.

Examples

julia> DateTime("2023-05-18T20:14:55.02")
2023-05-18T20:14:55.020

julia> Tempo.DateTime("2022-05-12")
2022-05-12T00:00:00.00

DateTime(seconds::Number)

Create a DateTime object given the number of seconds elapsed since J2000.

DateTime(d::Date, sec::Number)

Create a DateTime object given a Date and the number of seconds since midnight.

Examples

julia> d = Date(2023, 5, 18)
2023-05-18

julia> DateTime(d, 0)
2023-05-18T12:00:00.000

julia> DateTime(d, 1)
2023-05-18T12:00:01.000

See also

See also Date, Time and Epoch.

Tempo.DateTime — Method

DateTime(e::Epoch)

Construct a DateTime object from an Epoch.

Tempo.Epoch — Type

Epoch{S, T}

A type to represent Epoch-like data. Epochs are internally represented as seconds + fraction of seconds since a reference epoch, which is considered to be 2000-01-01T12:00:00, i.e. J2000.

Epoch(sec::Number, scale::AbstractTimeScale)
Epoch(sec::Number, scale::Type{<:AbstractTimeScale})
Epoch{S}(seconds::Number) where {S <: AbstractTimeScale}

Create an Epoch object from the number of seconds since J2000 with the timescale S.

Epoch(dt::DateTime, scale::AbstractTimeScale)
Epoch(dt::DateTime, scale::Type{<:AbstractTimeScale})

Create an Epoch object from a DateTime structure with timescale scale.

Epoch(str::AbstractString, scale::AbstractTimeScale)
Epoch(str::AbstractString)

Create an Epoch object from an ISO-formatted string. The timescale can either be specified as a second argument or written at the end of the string.

This constructor requires that the str is in the format:

ISO – yyyy-mm-ddTHH:MM:SS.ffff : assume J2000 as origin

J2000 – DDDD.ffff : parse Julian Date since J2000, in days

JD – JD DDDDDDDDD.ffffff : parse Julian Date, in days

MJD – MJD DDDDDDDDD.ffffff : parse a Modified Julian Date, in days

A TimeScale can be added at the end of the string, separated by a whitespace. If it is not declared, TDB will be used as a default timescale.

Examples

```julia-repl

standard ISO string

julia> Epoch("2050-01-01T12:35:15.0000 TT") 2050-01-01T12:35:14.9999 TT

standard ISO string (without scale)

julia> Epoch("2050-01-01T12:35:15.0000") 2050-01-01T12:35:14.9999 TDB

Parse Julian Dates

julia> Epoch("JD 2400000.5") 1858-11-17T00:00:00.0000 TDB

Parse Modified Julian Dates

julia> Epoch("MJD 51544.5") 2000-01-01T12:00:00.0000 TDB

Parse Julian Dates since J2000

julia> Epoch("12.0") 2000-01-13T12:00:00.0000 TDB

All Julian Date parsers allow timescales

julia> Epoch("12.0 TT") 2000-01-13T12:00:00.0000 TT

Tempo.EpochConversionError — Type

true

A type representing epoch conversion errors.

Tempo.GeocentricCoordinateTime — Type

GeocentricCoordinateTime <: AbstractTimeScale

A type representing the Geocentric Coordinate Time (TCG) time scale.

Tempo.GlobalPositioningSystemTime — Type

GlobalPositioningSystemTime <: AbstractTimeScale

A type representing the Global Positioning System Time (GPS) time scale.

Tempo.HighPrecisionBarycentricDynamicalTime — Type

HighPrecisionBarycentricDynamicalTime <: AbstractTimeScale

A type representing the High Precision Barycentric Dynamical Time (TDBH) time scale.

Tempo.InternationalAtomicTime — Type

InternationalAtomicTime <: AbstractTimeScale

A type representing the International Atomic Time (TAI) time scale.

Tempo.JulianDate — Type

JulianDate

A type representing the JulianDate (JD) epoch origin.

With this origin, Epoch reference is -4712-01-01T12:00.

Tempo.JulianDate2000 — Type

JulianDate2000

A type representing the JulianDate2000 (J2000) epoch origin.

With this origin, Epoch reference is 2000-01-01T12:00.

Tempo.Leapseconds — Type

Leapseconds{T}

Stores information about the leap seconds that have been added to Coordinated Universal Time (UTC).

Fields

lastupdate: a DatesDateTime object representing the date and time when the Leapseconds struct was last updated.

jd2000: a vector storing the Julian Date, in days since J2000, of each leap second.

leap: a vector storing the number of leap seconds at each corresponding entry of the jd2000 field.

Tempo.ModifiedJulianDate — Type

ModifiedJulianDate

A type representing the ModifiedJulianDate (MJD) epoch origin.

With this origin, Epoch reference is 1858-11-17T00:00.

Tempo.ModifiedJulianDate2000 — Type

ModifiedJulianDate2000

A type representing the ModifiedJulianDate2000 (MJD2000) epoch origin.

With this origin, Epoch reference is 2000-01-01T00:00.

Tempo.TerrestrialTime — Type

TerrestrialTime <: AbstractTimeScale

A type representing the Terrestrial Time (TT) time scale.

Tempo.Time — Type

Time{T}

A type representing the time of the day storing the hour, minute, seconds and fraction of seconds.

Time(hour::Int, minute::Int, second::Int, fraction::T) where {T <: Number}

Create a Time object of type T.

Time(hour::Int, minute::Int, second::Number)

Construct a Time object given the hour, minute and seconds. In this case, the seconds can either be an integer or a floating point number. The fraction of seconds will be computed under the hood.

Time(secondinday::Int, fraction::Number)
Time(secondinday::Number)

Create a Time object given the seconds of the day secondinday and/or the fraction of seconds.

Time(dt::DateTime)

Extract the Time object from a DateTime structure.

See also

See also Date and DateTime.

Tempo.TimeScaleNode — Type

TimeScaleNode{T} <: AbstractGraphNode

Define a timescale.

Fields

name – timescale name

id – timescale identification number (ID)

parentid – ID of the parent timescale

ffp – offest function from the parent timescale

ftp – offset function to the parent timescale

Tempo.TimeSystem — Type

TimeSystem{T}

A TimeSystem object manages a collection of default and user-defined TimeScaleNode objects, enabling efficient time transformations between them. It leverages a MappedDiGraph to keep track of the relationships between the timescales.

TimeSystem{T}()

Create a empty TimeSystem object with datatype T.

Examples

julia> ts = TimeSystem{Float64}();

julia> @timescale TSA 100 TimeScaleA

julia> @timescale TSB 200 TimeScaleB

julia> add_timescale!(ts, TSA)

julia> offset_tsa2tsb(seconds) = 1.0

julia> offset_tsb2tsa(seconds) = -1.0

julia> add_timescale!(ts, TSB, offset_tsa2tsb; parent=TSA, ftp=offset_tsb2tsa)

See also

See also @timescale and add_timescale!.

Tempo.UniversalTime — Type

UniversalTime <: AbstractTimeScale

A type representing the Universal Time (UT1) time scale.

Types

Base.convert — Method

convert(to::S2, e::Epoch{S1}; system::TimeSystem=TIMESCALES)

Convert Epoch with timescale S1 to S2. Allows to use the default TimeSystem or a custom constructed one.

Tempo.add_timescale! — Method

add_timescale!(system::TimeSystem, scale::AbstractTimeScale, ffp::Function; ftp, parent)

Add scale as a timescale to system. A custom function ffp providing the time offset, in seconds, between the parent scale and the current scale must be provided by the user.

The parent and ffp arguments are unneeded only for the root timescale. If the user wishes to add a scale to a non-empty timesystem, this argument becomes mandatory.

The input functions must accept only the seconds in the parent scale as argument and must return a single numerical output. An optional function ftp, with a similar interface, returning the offset from the current to the parent scale may also be provided.

Note

If ftp is not provided, the reverse timescale transformation will not be possible.

Examples

```julia-repl julia> SYSTEM = TimeSystem{Float64}();

julia> @timescale RTS 102 RootTimeScale

julia> @timescale CTS 103 ChildTimeScale

julia> roottochild(x::Number) = 13.3;

julia> childtoroot(x::Number) = -13.3;

julia> add_timescale!(SYSTEM, RTS)

julia> addtimescale!(SYSTEM, CTS, roottochild; parent=RTS, ftp=childto_root)

See also

See also @timescale, TimeSystem and apply_offsets.

Tempo.add_timescale! — Method

add_timescale!(s::TimeSystem, ts::TimeScaleNode)

Register a new node in the TimeSystem.

Warning

This is a low-level function and should not be called by the user.

Tempo.cal2jd — Method

cal2jd(d::Date)

Convert Gregorian calendar Date to a Julian Date, in days.

Outputs

j2000 – J2000 zero point: always 2451545

d – J2000 Date for 12 hrs

Tempo.cal2jd — Method

cal2jd(year::Integer, month::Integer, day::Integer)

This function converts a given date in the Gregorian calendar (year, month, day) to the corresponding two-parts Julian Date. The first part is the DJ2000, while the second output is the number of days since J2000.

The year must be greater than 1583, and the month must be between 1 and 12. The day must also be valid, taking into account whether the year is a leap year. If the input year or month or day are invalid, a DomainError is thrown.

Examples

julia> Tempo.cal2jd(2021, 1, 1)
(2.4000005e6, 59215.0)

julia> Tempo.cal2jd(2022, 2, 28)
(2.4000005e6, 59638.0)

julia> Tempo.cal2jd(2019, 2, 29)
ERROR: DomainError with 29:
the day shall be between 1 and 28.

References

Seidelmann P. K., (1992), Explanatory Supplement to the Astronomical Almanac, University Science Books, Section 12.92 (p604).

Klein, A., (2006), A Generalized Kahan-Babuska-Summation-Algorithm. Computing, 76, 279-293, Section 3.

ERFA software library

Tempo.calhms2jd — Method

calhms2jd(year, month, day, hour, minute, seconds)

Convert Gregorian Calendar date and time to a two-parts Julian Date. The first part is the DJ2000, while the second output is the number of days since J2000.

Examples

julia> Tempo.calhms2jd(2000, 1, 1, 12, 0, 0)
(2.451545e6, 0.0)

julia> Tempo.calhms2jd(2022, 1, 1, 0, 0, 0)
(2.451545e6, 8035.5)

Tempo.day — Method

day(d::DateTime)

Get day associated to a DateTime type.

Tempo.day — Method

day(d::Date)

Get day associated to a Date.

Tempo.fd2hms — Method

fd2hms(fd::Number)

Convert the day fraction fd to hour, minute and seconds.

Tempo.fd2hmsf — Method

fd2hmsf(fd::Number)

Convert the day fraction fd to hour, minute, second and fraction of seconds.

Examples

julia> Tempo.fd2hms(0.5)
(12, 0, 0.0)

Tempo.find_day — Method

find_day(dayinyear::Integer, month::Integer, isleap::Bool)

Find the day of the month from the day in the year and the month, depending on whether the year is leap or not.

Tempo.find_dayinyear — Method

find_dayinyear(d::Date)

Find the day in the year.

Tempo.find_dayinyear — Method

find_dayinyear(month::Integer, day::Integer, isleap::Bool)

Find the day of the year given the month, the day of the month and whether the year is leap or not.

Tempo.find_month — Method

find_month(dayinyear::Integer, isleap::Bool)

Find the month from the day of the year, depending on whether the year is leap or not.

Tempo.find_year — Method

find_year(d::Integer)

Return the Gregorian year associated to the given Julian Date day d since J2000.

Tempo.fraction_of_day — Method

fraction_of_day(t::Time)
hms2fd(t::Time)

Find the fraction of the day.

Example

julia> t = Time(12, 30, 40.3424)
12:30:40.3423

julia> Tempo.fraction_of_day(t)
0.5213002592592593

Tempo.fraction_of_second — Method

fraction_of_second(d::DateTime)

Get the fraction of seconds associated to a DateTime object.

Tempo.fraction_of_second — Method

fraction_of_second(t::Time)

Find the fraction of seconds.

Example

julia> t = Time(12, 30, 40.3424)
12:30:40.3423

julia> Tempo.fraction_of_second(t)
0.3423999999999978

Tempo.get_leapseconds — Method

get_leapseconds()

Parse leapseconds data and return a Leapseconds type.

Note

The leapsecond kernel is retrieved from the artifacts of this package. This artifact will be updated whenever a new leapsecond is added.

Tempo.hms2fd — Method

hms2fd(hour::Integer, minute::Integer, second::Number)

Convert hours, minutes and seconds to day fraction.

Examples

```julia-repl julia> Tempo.hms2fd(12, 0.0, 0.0) 0.5

Tempo.hour — Method

hour(d::DateTime)

Get hour associated to a DateTime type.

Tempo.hour — Method

hour(t::Time)

Get the current hour.

Tempo.isleapyear — Method

isleapyear(d::Date)

True if Date is within a leap year.

Tempo.isleapyear — Method

isleapyear(year::Integer)

Return true if the given Gregorian year is leap.

Tempo.j2000 — Method

j2000(jd)
j2000(jd1, jd2)

Convert Julian Date in days since J2000

Tempo.j2000 — Method

j2000(dt::DateTime)

Convert a DateTime dt in Julian days since J2000.

Tempo.j2000 — Method

j2000(d::Date)

Convert Gregorian calendar date Date to a Julian Date since J2000, in days.

Tempo.j2000 — Method

j2000(e::Epoch)

Convert Epoch in Julian Date days since J2000.

Tempo.j2000c — Method

j2000c(jd)
j2000c(jd1, jd2)

Convert Julian Date (in days) to Julian Centuries

Tempo.j2000c — Method

j2000c(dt::DateTime)

Convert a DateTime dt in a Julian Date since J2000, in centuries.

Tempo.j2000c — Method

j2000c(e::Epoch)

Convert Epoch in Julian Date centuries since J2000.

Tempo.j2000s — Method

j2000s(jd)
j2000s(jd1, jd2)

Convert Julian Date (in days) in seconds past J2000

Tempo.j2000s — Method

j2000s(dt::DateTime)

Convert a DateTime dt to seconds since J2000.

Tempo.j2000s — Method

j2000s(e::Epoch)

Convert Epoch in Julian Date seconds since J2000.

Tempo.jd2cal — Method

jd2cal(dj1::Number, dj2::Number)

This function converts a given Julian Date (JD) to a Gregorian calendar date (year, month, day, and fraction of a day).

Examples

julia> Tempo.jd2cal(DJ2000, 0.0)
(2000, 1, 1, 0.5)

julia> Tempo.jd2cal(DJ2000, 365.5)
(2001, 1, 1, 0.0)

julia> Tempo.jd2cal(DJ2000 + 365, 0.5)
(2001, 1, 1, 0.0)

Note

The Julian Date is apportioned in any convenient way between the arguments dj1 and dj2. For example, JD = 2450123.7 could be expressed in any of these ways, among others:

dj1	dj2

2450123.7	0.0	(JD method)

2451545.0	-1421.3	(J2000 method)

2400000.5	50123.2	(MJD method)

2450123.5	0.2	(date & time method)

Warning

The earliest valid date is -68569.5 (-4713 Jan 1). The largest value accepted is 1e9.

References

Seidelmann P. K., (1992), Explanatory Supplement to the Astronomical Almanac, University Science Books, Section 12.92 (p604).

Klein, A., (2006), A Generalized Kahan-Babuska-Summation-Algorithm. Computing, 76, 279-293, Section 3.

ERFA software library

Tempo.jd2calhms — Method

jd2calhms(dj1::Number, dj2::Number)

Convert a two-parts Julian Date to Gregorian year, month, day, hour, minute, seconds. See jd2cal for more information on the Julian Date composition.

julia> Tempo.jd2calhms(DJ2000, 0.0)
(2000, 1, 1, 12, 0, 0.0)

julia> Tempo.jd2calhms(DJ2000 + 1, 0.25)
(2000, 1, 2, 18, 0, 0.0)

julia> Tempo.jd2calhms(DJ2000, 1.25)
(2000, 1, 2, 18, 0, 0.0)

Tempo.lastj2000dayofyear — Method

lastj2000dayofyear(year::Integer)

Tempo.leapseconds — Method

leapseconds(jd2000::Number)

For a given UTC date, in Julian days since J2000, calculate Delta(AT) = TAI - UTC.

Tempo.minute — Method

minute(d::DateTime)

Get minute associated to a DateTime type.

Tempo.minute — Method

minute(t::Time)

Get the current minute.

Tempo.month — Method

month(d::DateTime)

Get month associated to a DateTime type.

Tempo.month — Method

month(d::Date)

Get month associated to a Date.

Tempo.offset — Method

offset(::JulianDate2000)

Offset in days to shift J2000 epochs (with origin at 2000-01-01T12:00) to JulianDate2000 (with origin at 2000-01-01T12:00)

Tempo.offset — Method

offset(::JulianDate)

Offset in days to shift J2000 epochs (with origin at 2000-01-01T12:00) to JulianDate (with origin at -4712-01-01T12:00)

Tempo.offset — Method

offset(::ModifiedJulianDate2000)

Offset in days to shift J2000 epochs (with origin at 2000-01-01T12:00) to ModifiedJulianDate2000 (with origin at 2000-01-01T00:00)

Tempo.offset — Method

offset(::ModifiedJulianDate)

Offset in days to shift J2000 epochs (with origin at 2000-01-01T12:00) to ModifiedJulianDate (with origin at 1858-11-17T00:00)

Tempo.offset_gps2tai — Method

offset_gps2tai(seconds)

Return the fixed offset between GPS and TAI in seconds.

Tempo.offset_tai2gps — Method

offset_tai2gps(seconds)

Return the fixed offset between TAI and GPS in seconds.

Tempo.offset_tai2tt — Method

offset_tai2tt(seconds)

Return the fixed offset between TAI and TT in seconds.

Tempo.offset_tai2utc — Method

offset_tai2utc(seconds)

Return the offset between TAI and UTC in seconds.

Tempo.offset_tcb2tdb — Method

offset_tcb2tdb(seconds)

Return the linear offset between TCB and TDB in seconds.

Tempo.offset_tcg2tt — Method

offset_tcg2tt(seconds)

Return the linear offset between TCG and TT in seconds.

Tempo.offset_tdb2tcb — Method

offset_tdb2tcb(seconds)

Return the linear offset between TDB and TCB in seconds.

Tempo.offset_tdb2tt — Method

offset_tdb2tt(seconds)

Return the offset between TDB and TT in seconds. This routine is accurate to ~40 microseconds over the interval 1900-2100.

Note

An accurate transformation between TDB and TT depends on the trajectory of the observer. For two observers fixed on Earth's surface the quantity TDB-TT can differ by as much as ~4 microseconds.

References

https://www.cv.nrao.edu/~rfisher/Ephemerides/times.html#TDB

Issue #26

Tempo.offset_tt2tai — Method

offset_tt2tai(seconds)

Return the fixed offset between TT and TAI in seconds.

Tempo.offset_tt2tcg — Method

offset_tt2tcg(seconds)

Return the linear offset between TT and TCG in seconds.

Tempo.offset_tt2tdb — Method

offset_tt2tdb(seconds)

Return the offset between TT and TDB in seconds. This routine is accurate to ~40 microseconds over the interval 1900-2100.

Note

An accurate transformation between TDB and TT depends on the trajectory of the observer. For two observers fixed on Earth's surface the quantity TDB-TT can differ by as much as ~4 microseconds.

References

https://www.cv.nrao.edu/~rfisher/Ephemerides/times.html#TDB

Issue #26

Tempo.offset_tt2tdbh — Method

offset_tt2tdbh(seconds)

Return the offset between TT and TDBH in seconds.

The maximum error in using the above formula is about 10 µs from 1600 to 2200. For even more precise applications, the series expansion by Harada & Fukushima (2003) is recommended.

References

The IAU Resolutions on Astronomical Reference Systems, Time Scales, and Earth Rotation Models, United States Naval Observatory, https://arxiv.org/pdf/astro-ph/0602086.pdf

Tempo.offset_utc2tai — Method

offset_utc2tai(seconds)

Return the offset between UTC and TAI in seconds.

Tempo.parse_iso — Method

parse_iso(s::AbstractString)

Parse an ISO datetime string, in the format YYYY-MM-DDThh:mm:ss.ffffffff and return a Tuple containing the year, month, day , hour, minute, second and milliseconds. If the string format is not recognised as ISO, an ArgumentError is thrown.

Tempo.second — Method

second(::Type{<:AbstractFloat}, t::Time)
second(::Type{<:Integer}, t::Time)
second(t::Time)

Get the current second.

Tempo.second — Method

second(::Type{<:AbstractFloat}, t::Time)
second(d::DateTime)

Get the seconds associated to a DateTime type. If a floating-point type is given as first argument, the returned value will also account for the fraction of seconds.

Tempo.second_in_day — Method

second_in_day(t::Time)

Find the second in the day.

Example

julia> t = Time(12, 30, 40.3424)
12:30:40.3423

julia> Tempo.second_in_day(t)
45040.3424

Tempo.tai2utc — Method

tai2utc(tai1, tai2)

Transform a 2-part (quasi) Julian Date, in days, in International Atomic Time, TAI to a 2-part Julian Date in the Coordinated Universal Time, UTC, scale.

Note

tai1 + tai2 is Julian Date, apportioned in any convenient way between the two arguments, for example such that tai1 is the Julian Day Number and tai2 is the fraction of a day. The returned utc1 and utc2 form an analogous pair.

Note

JD cannot unambiguously represent UTC during a leap second unless special measures are taken. The convention in the present function is that the JD day represents UTC days whether the length is 86399, 86400 or 86401 SI seconds.

References

Seidelmann P. K., (1992), Explanatory Supplement to the Astronomical Almanac, University Science Books, Section 12.92 (p604).

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

ERFA software library

Tempo.timescale — Method

timescale(ep::Epoch)

Epoch timescale.

Tempo.timescale_alias — Method

timescale_alias(scale::AbstractTimeScale)

Return the ID associated to scale.

Tempo.timescale_id — Method

timescale_id(scale::AbstractTimeScale)

Return the ID of scale.

Tempo.timescale_name — Method

timescale_name(scale::AbstractTimeScale)

Return the name of scale.

Tempo.utc2tai — Method

utc2tai(utc1, utc2)

Transform a 2-part (quasi) Julian Date, in days, in Coordinate Universal Time, UTC to a 2-part Julian Date in the International Atomic Time, TAI scale.

Note

utc1 + utc2 is quasi Julian Date (see Note 2), apportioned in any convenient way between the two arguments, for example such that utc1 is the Julian Day Number and utc2 is the fraction of a day.

Note

JD cannot unambiguously represent UTC during a leap second unless special measures are taken. The convention in the present function is that the JD day represents UTC days whether the length is 86399, 86400 or 86401 SI seconds.

References

Seidelmann P. K., (1992), Explanatory Supplement to the Astronomical Almanac, University Science Books, Section 12.92 (p604).

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

ERFA software library

Tempo.value — Method

value(ep::Epoch)

Full Epoch value.

Tempo.year — Method

year(d::DateTime)

Get year associated to a DateTime type.

Tempo.year — Method

year(d::Date)

Get year associated to a Date.

Macros

Tempo.@timescale — Macro

@timescale(name, id, type)

Create a new timescale instance to alias the given id. This macro creates an AbstractTimeScale subtype and its singleton instance called name. Its type is obtained by appending TimeScale to name if it was not provided.

Examples

julia> @timescale NTS 100 NewTimeScale 

julia> typeof(NTS)
NewTimeScale 

julia> timescale_alias(NTS)
100

julia> @timescale TBH 200

julia> typeof(TBH)
TBHTimeScale

julia> timescale_alias(TBH)
200

See also

See also timescale_alias and add_timescale!.